f32->f64
Eventually these might be made generic, but at this time f64 is more useful for
scientific work.

+++ src/macros.rs
@@ -13,7 +13,7 @@ macro_rules! vec
 	( $($e: expr),+) =>
 	{
 		Vector::new([$(
-				($e) as f32,
+				($e) as f64,
 		)+])
 	}
 }
@@ -25,7 +25,7 @@ macro_rules! mat
 	{
 		Matrix::new([$(
 			&[$(
-				($e) as f32,
+				($e) as f64,
 			)+],
 		)+])
 	}
