+++ src/matrix/column_accessor.rs
@@ -40,12 +40,12 @@ impl<'l,
 VectorGet for
 ColumnAccessor<T>
 {
-	unsafe fn unsafe_get(&self, idx: uint) -> f32
+	unsafe fn unsafe_get(&self, idx: uint) -> f64
 	{
 		self.base.unsafe_get(idx, self.col)
 	}
 
-	fn get(&self, idx: uint) -> f32
+	fn get(&self, idx: uint) -> f64
 	{
 		assert!(idx < self.base.ncol());
 		unsafe
